requirements/prod.txt
Django==2.0.6
django-allauth==0.36.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-oauth-toolkit==1.2.0
django-cors-middleware==1.3.1
django-rest-swagger==2.2.0
djangorestframework-filters==0.10.2
django-role-permissions==2.2.0
africastalking==1.1.4
